refactor: move isMentor selector to register route
